Revert "Kill virtualenv on the fish prompt again"
This reverts commit 41b353c3152fdd57e773c1e3ad2437faf54d8c51.

I can't decide!!!!!! Sometimes I end up with the virtualenv twice in the prompt, sometimes
I end up with no indication of the active virtualenv at all. Let's duplicate for now and
maybe try to come up later with a way of adding this part of the prompt only if the name
of the virtualenv is not in the current prompt already. Maybe that can be done.